+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.2.0 under development
 
-- New #4: Add `package-authors-default-source` setting to `PackageAuthors` fact that controls the default author source
+- New #4, #6: Add `package-authors-default-source` setting to `PackageAuthors` fact that controls the default author source
   when authors are not provided explicitly and absent in `composer.json`: take from user (`user`, default) or use an
   empty list (`empty`).